rename summary lens to markdown lens
