+++ cmd/deck/main.go
@@ -94,7 +94,7 @@ import (
 	_ "sigs.k8s.io/prow/pkg/spyglass/lenses/metadata"
 	_ "sigs.k8s.io/prow/pkg/spyglass/lenses/podinfo"
 	_ "sigs.k8s.io/prow/pkg/spyglass/lenses/restcoverage"
-	_ "sigs.k8s.io/prow/pkg/spyglass/lenses/summary"
+	_ "sigs.k8s.io/prow/pkg/spyglass/lenses/markdown"
 )
 
 // Omittable ProwJob fields.
